Aspect feature
Added a button in toolbar for setting aspect ratios, I have added the three standard ratios 16:9 landscape, 9:16 portrait, 1:1 insta post. The code is properly commented such that any other contributor can add more ratios easily if they wish to.
Added proper styling of the button and dropdown even in dark mode,
The respective buttons change the existing canvas to the preset size and retain the bgcolor and pattern of the previous canvas when undergoing change.
Please note that the increase and decrease height do not work on preset aspect ratios because then the purpose of the ratios will not remain.
I have also added this feature to the guidelines panel.